+++ Engine/Pad.h
@@ -9,15 +9,15 @@ class Pad
 private:
 	Vec2 pos;
 	Color c = Colors::White;
-	int width;
-	int height;
-	int speed; //can only move on x axis
+	float width;
+	float height;
+	float speed; //can only move on x axis
 	Rect rect;
 public:
 	Pad() = default;
-	Pad(Vec2 pos, int width, int height, int speed);
+	Pad(Vec2 pos, float width, float height, float speed);
 	void Draw(Graphics& gfx);
-	void Move(int delta_x, float dt);
+	void Move(float delta_x, float dt);
 	void KeepInBounds(Rect& walls);
 	void DoBallCollision(Ball& ball);
