+++ Engine/Vec2.cpp
@@ -1,6 +1,6 @@
 #include "Vec2.h"
 
-Vec2::Vec2(int x, int y)
+Vec2::Vec2(float x, float y)
 	:
 	x(x),
 	y(y)
@@ -17,12 +17,12 @@ Vec2& Vec2::operator+=(Vec2& rhs)
 	return *this = *this + rhs;
 }
 
-Vec2 Vec2::operator*(int rhs) const
+Vec2 Vec2::operator*(float rhs) const
 {
 	return Vec2(x*rhs, y*rhs);
 }
 
-Vec2& Vec2::operator*=(int rhs)
+Vec2& Vec2::operator*=(float rhs)
 {
 	return *this = *this * rhs;
 }
